McGillHackaton-main/requirements.txt
adagio==0.2.6
alembic==1.13.2
annotated-types==0.7.0
anyio==4.4.0
appdirs==1.4.4
appnope==0.1.4
arch==7.0.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
astropy==6.1.4
astropy-iers-data==0.2024.9.30.0.32.59
asttokens==2.4.1
async-lru==2.0.4
attrs==24.2.0
autots==0.6.15
babel==2.16.0
beautifulsoup4==4.12.3
bleach==6.1.0
blinker==1.8.2
catboost==1.2.7
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.3.2
clarabel==0.9.0
click==8.1.7
cloudpickle==3.0.0
colorlog==6.8.2
comm==0.2.2
contourpy==1.3.0
coreforecast==0.0.12
cramjam==2.8.3
cvxpy==1.5.3
cycler==0.12.1
Cython==3.0.11
dash==2.18.1
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-table==5.0.0
debugpy==1.8.5
decorator==5.1.1
defusedxml==0.7.1
dtaidistance==2.3.12
dtw-python==1.5.3
ecos==2.0.14
et-xmlfile==1.1.0
executing==2.1.0
fastjsonschema==2.20.0
fastparquet==2024.5.0
Flask==3.0.3
fonttools==4.53.1
fqdn==1.5.1
frozendict==2.4.4
fs==2.4.16
fsspec==2024.9.0
fugue==0.9.1
gluonts==0.15.1
graphviz==0.20.3
h11==0.14.0
hmmlearn==0.3.2
holidays==0.57
html5lib==1.1
httpcore==1.0.5
httpx==0.27.2
idna==3.8
importlib_metadata==8.5.0
ipykernel==6.29.5
ipython==8.27.0
isoduration==20.11.0
itsdangerous==2.2.0
jedi==0.19.1
Jinja2==3.1.4
joblib==1.4.2
json5==0.9.25
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
jupyter-events==0.10.0
jupyter-lsp==2.2.5
jupyter_client==8.6.2
jupyter_core==5.7.2
jupyter_server==2.14.2
jupyter_server_terminals==0.5.3
jupyterlab==4.2.5
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.3
kaleido==0.2.1
kiwisolver==1.4.7
lazy_loader==0.4
lightgbm==4.5.0
llvmlite==0.43.0
lxml==5.3.0
Mako==1.3.5
MarkupSafe==2.1.5
matplotlib==3.9.2
matplotlib-inline==0.1.7
mistune==3.0.2
mne==1.8.0
multitasking==0.0.11
nbclient==0.10.0
nbconvert==7.16.4
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.3
nfoursid==1.0.1
notebook==7.2.2
notebook_shim==0.2.4
numba==0.60.0
numpy==1.26.4
openpyxl==3.1.5
optuna==4.0.0
osqp==0.6.7.post1
overrides==7.7.0
packaging==23.2
pandas==2.2.2
pandocfilters==1.5.1
parso==0.8.4
patsy==0.5.6
peewee==3.17.6
pexpect==4.9.0
pillow==10.4.0
platformdirs==4.3.2
plotly==5.24.1
pmdarima==2.0.4
polars==1.7.1
pooch==1.8.2
prometheus_client==0.20.0
prompt_toolkit==3.0.47
psutil==6.0.0
psycopg2-binary==2.9.9
ptyprocess==0.7.0
pure_eval==0.2.3
pyaml==24.7.0
pyarrow==17.0.0
pybind11==2.13.6
pycatch22==0.4.5
pycparser==2.22
pydantic==2.9.2
pydantic_core==2.23.4
pyerfa==2.0.1.4
Pygments==2.18.0
pyod==2.0.2
pyparsing==3.1.4
python-dateutil==2.9.0.post0
python-json-logger==2.0.7
pytz==2024.1
PyYAML==6.0.2
pyzmq==26.2.0
qdldl==0.1.7.post4
quantstats-lumi==0.3.3
referencing==0.35.1
requests==2.32.3
retrying==1.3.4
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
Riskfolio-Lib==6.2.3
rpds-py==0.20.0
scikit-base==0.8.3
scikit-learn==1.5.1
scikit-optimize==0.10.2
scikit-posthocs==0.9.0
scipy==1.12.0
scs==3.2.7
seaborn==0.13.2
seasonal==0.3.1
Send2Trash==1.8.3
setuptools==74.1.2
shap==0.46.0
six==1.16.0
skfolio==0.4.1
skforecast==0.13.0
skpro==2.5.1
sktime==0.33.0
slicer==0.0.8
sniffio==1.3.1
soupsieve==2.6
SQLAlchemy==2.0.34
stack-data==0.6.3
statsforecast==1.7.7.1
statsmodels==0.14.3
tabulate==0.9.0
tbats==1.1.3
tenacity==9.0.0
terminado==0.18.1
threadpoolctl==3.5.0
tinycss2==1.3.0
toolz==0.12.1
tornado==6.4.1
tqdm==4.66.5
traitlets==5.14.3
triad==0.9.8
types-python-dateutil==2.9.0.20240906
typing_extensions==4.12.2
tzdata==2024.1
u8darts==0.30.0
ujson==5.10.0
uri-template==1.3.0
urllib3==2.2.2
utilsforecast==0.2.5
wcwidth==0.2.13
webcolors==24.8.0
webencodings==0.5.1
websocket-client==1.8.0
Werkzeug==3.0.4
wrds==3.2.0
xarray==2024.9.0
xgboost==2.1.1
XlsxWriter==3.2.0
yfinance==0.2.44
zipp==3.20.2
